release: 0.6.94
